PLAT-729 - refactor; move all legwork to util.py
