+++ deploy.py
@@ -25,7 +25,6 @@ import sys
 import shutil
 import pdb
 import json
-from tempfile import NamedTemporaryFile
 
 
 logger = logging.getLogger(__name__)
@@ -104,6 +103,13 @@ class Deployment:
         env['AWS_SESSION_TOKEN'] = aws_session_token
         env['AWS_DEFAULT_REGION'] = self.metadata['REGION']
 
+        # process secrets
+        if "CREDSTASH" in self.metadata:
+            secrets = util.Credstash().process(self.metadata["CREDSTASH"], self.metadata['TEAM'],
+                                               self.component_name, self.environment, env)
+        else:
+            secrets = None
+
         print("Preparing S3 bucket for terragrunt...")
         s3_bucket_name = self.terragrunt_s3_bucket_name()
         self.s3_bucket_prep(s3_bucket_name)
@@ -115,73 +121,14 @@ class Deployment:
         # get all the relevant modules
         check_call("terraform get infra", env=env, shell=True)
 
-        # process secrets
-        decrypted_secrets_file = None
-        if "CREDSTASH" in self.metadata:
-            if self._secrets(self.metadata["CREDSTASH"]):
-                decrypted_secrets_file = self._generate_decrypted_credentials(self.metadata["TEAM"], self.component_name, self.environment, env)
-
         self.terragrunt('plan', self.environment, self.ecr_image_name, self.component_name, self.metadata['REGION'],
-                        self.metadata['TEAM'], self.version, decrypted_secrets_file, env)
+                        self.metadata['TEAM'], self.version, secrets, env)
         if self.plan is False:
             self.terragrunt('apply', self.environment, self.ecr_image_name, self.component_name, self.metadata['REGION'],
-                            self.metadata['TEAM'], self.version, decrypted_secrets_file, env)
+                            self.metadata['TEAM'], self.version, secrets, env)
 
         # clean up all irrelevant files
         self.cleanup()
-
-    def _secrets(self, CREDSTASH):
-        """
-        Helper method to check whether we need to process secrets or not
-
-        Params:
-            CREDSTASH: string either true/false from service.json indicating
-                       whether to fetch secrets using credstash
-        Returns:
-            bool: True is it exist, False if not
-        """
-
-        if CREDSTASH == "true":
-            return True
-        else:
-            return False
-
-    def _credstash_getall(self, team, exec_env):
-        """
-        Get all secrets for a specific team (we filter them later)
-
-        Params:
-            team: team; used to differentiate which KMS master key to use
-            exec_env: used by subprocess call
-
-        Returns:
-            string: result credstash got back the "vault"
-        """
-        s = check_output(["credstash", "-t", "credstash-%s" % team,
-                                       "getall"], env=exec_env)
-        return str(s)
-
-    def _generate_decrypted_credentials(self, team, component, env, exec_env):
-        """
-        Params:
-            team: needed by _credstash_get
-            component: needed to filter the result
-            env: needed to filter the result
-            exec_env: needed by _credstash_get
-        Returns:
-            file: file-location with the resulting json
-        """
-        feed = json.loads(self._credstash_getall(team, exec_env))
-        secrets = {}
-
-        prefix = "deploy.{environment}.{component}.".format(environment=env,component=component)
-        for key in feed.keys():
-            if key.startswith(prefix):
-                secrets[str(key)[len(prefix):]] = feed[key]
-
-        f = NamedTemporaryFile(delete=False)
-        f.write(json.dumps({"secrets": secrets}))
-        return f.name
 
     def get_aws(self):
         """
